(Refact) detach logout
